EZEE-3065: moved test to SearchService dir, corrected strict type
